+++ src/avfilter/avfilter.rs
@@ -45,7 +45,7 @@ impl AVFilterContext {
             )
         }
         .upgrade()
-        .map_err(|_| RsmpegError::SetPropertyError)?;
+        .map_err(RsmpegError::SetPropertyError)?;
         Ok(())
     }
 
@@ -65,7 +65,7 @@ impl AVFilterContext {
 
         unsafe { ffi::av_buffersrc_add_frame_flags(self.as_mut_ptr(), frame_ptr, flags) }
             .upgrade()
-            .map_err(|_| RsmpegError::BufferSrcAddFrameError)?;
+            .map_err(RsmpegError::BufferSrcAddFrameError)?;
         Ok(())
     }
 
@@ -82,7 +82,7 @@ impl AVFilterContext {
             Ok(_) => Ok(frame),
             Err(AVERROR_EAGAIN) => Err(RsmpegError::BufferSinkDrainError),
             Err(ffi::AVERROR_EOF) => Err(RsmpegError::BufferSinkEofError),
-            Err(_) => Err(RsmpegError::BufferSinkGetFrameError),
+            Err(err) => Err(RsmpegError::BufferSinkGetFrameError(err)),
         }
     }
 }
@@ -132,11 +132,17 @@ impl AVFilterGraph {
     pub fn parse_ptr(
         &self,
         filter_spec: &CStr,
-        mut inputs: AVFilterInOut,
-        mut outputs: AVFilterInOut,
+        mut inputs: Option<AVFilterInOut>,
+        mut outputs: Option<AVFilterInOut>,
     ) -> Result<(Option<AVFilterInOut>, Option<AVFilterInOut>)> {
-        let mut inputs_new = inputs.as_mut_ptr();
-        let mut outputs_new = outputs.as_mut_ptr();
+        let mut inputs_new = inputs
+            .as_mut()
+            .map(|x| x.as_mut_ptr())
+            .unwrap_or(ptr::null_mut());
+        let mut outputs_new = outputs
+            .as_mut()
+            .map(|x| x.as_mut_ptr())
+            .unwrap_or(ptr::null_mut());
 
         // FFmpeg `avfilter_graph_parse*`'s documentation states:
         //
@@ -158,10 +164,10 @@ impl AVFilterGraph {
         .upgrade()?;
 
         // If no error, inputs and outputs pointer are dangling, manually erase
-        // them without dropping. Do this because we need to drop inputs and
-        // outputs on the error path.
-        let _ = inputs.into_raw();
-        let _ = outputs.into_raw();
+        // them *without* dropping. Do this because we need to drop inputs and
+        // outputs on the error path, but don't drop them on normal path.
+        let _ = inputs.map(|x| x.into_raw());
+        let _ = outputs.map(|x| x.into_raw());
 
         // ATTENTION: TODO here we didn't bind the AVFilterInOut to the lifetime of the AVFilterGraph
         let new_inputs = inputs_new
@@ -179,6 +185,15 @@ impl AVFilterGraph {
         unsafe { ffi::avfilter_graph_config(self.as_ptr() as *mut _, ptr::null_mut()) }
             .upgrade()?;
         Ok(())
+    }
+
+    /// Get a filter instance identified by instance name from graph.
+    pub fn get_filter(&mut self, name: &CStr) -> Option<AVFilterContextMut> {
+        unsafe {
+            ffi::avfilter_graph_get_filter(self.as_mut_ptr(), name.as_ptr())
+                .upgrade()
+                .map(|raw| AVFilterContextMut::from_raw(raw))
+        }
     }
 }
 
@@ -208,7 +223,7 @@ impl<'graph> AVFilterGraph {
             )
         }
         .upgrade()
-        .map_err(|_| RsmpegError::CreateFilterError)?;
+        .map_err(RsmpegError::CreateFilterError)?;
 
         let filter_context = NonNull::new(filter_context).unwrap();